test(memory share): lend/relinquish/reclaim multiple borrowers
The test aims at validating that a memory region can be lent to multiple
borrowers, that they can access the memory, and give it back to the
lender.
As such, within the same context validate the borrowers can relinquish
access, and that the lender can reclaim its access to the memory region.

Change-Id: Ic8ea9a69bd0478ef03176a9ee5249d18019b9349
Signed-off-by: J-Alves <joao.alves@arm.com>