fix: support first version of SP package
To comply with FF-A v1.1 EAC0 boot protocol specification, the sp pkg
had been refactored to allocate space for the boot info blob.
The change broke the support for the first version of the sp pkg.
This patch refactors the sp_pkg module for hafnium to support both
versions and uses to the SP package.

Change-Id: If5a6382272fa0f2cce0e13fce4f80bedff9a6517
Signed-off-by: J-Alves <joao.alves@arm.com>