fix(boot protocol): check if 'gp-register-num' set
The manifests field 'gp-register-num' configures the register in which
the partition expects to receive the address of the boot info blob.
The section 5.4.3 of the FF-A v1.1 EAC0 specification states the
presence of this field should be indicative to the partition manager
that it should produce boot information. This patch skips processing
the boot info blob if the referred is not present in the manifest.

Change-Id: I7ced1414217c358fbcc675dc4ab5d3550d110bb6
Signed-off-by: J-Alves <joao.alves@arm.com>