feat(memory share): fragmented memory region with multiple borrowers
Refactored handling of ffa_memory_retrieve_continue to comply with
a fragmented memory share operation. This happens when the description
of all the constituents to be shared do not fit the RX/TX buffers.
In this case, after the first successful call to FFA_MEMORY_RETRIEVE_REQ
the SPMC should eret FFA_MEMORY_RETRIEVE_RESP along with: the respective
handle, the retrieved fragment length and total length.
When the fragment length is smaller than the total length, this means
there are more fragments to retrieve.
The borrower should then continue the retrieve operation using the
FFA_MEM_FRAG_RX ABI.

Change-Id: I1f475c1e6ec9c2abb21f3883abcb41a86d0e3f5d
Signed-off-by: J-Alves <joao.alves@arm.com>