test: SP preemption by non-secure interrupt
This test arms a timer in the PVM to expire later, then send a direct
request to a SP to query it to run a long busy loop.
The non-secure timer (physical) interrupt fires in the secure world
while a SP vCPU runs, traps to the SPMC, freezes the SP and returns
to the normal world through the FFA_INTERRUPT ABI.
Managed exit operation is not involved for S-EL1 partitions.
The PVM is resumed, traps to IRQ vector, handles the timer interrupt and
observes FFA_INTERRUPT status interface as a return from the direct
request. The PVM then resumes again the SP through FFA_RUN. The SP
completes the busy loop and terminates by sending a direct response
as a success.

Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Change-Id: I7e7a1fa049dcf6292fe18504b7ac898f146f2590