test: increase SPMC test coverage
The SPMC test coverage is extended. This patch adds a test setup
invoked in 'test_spmc.sh', based on 'primary_with_secondaries'
artifacts.
This setup deploys:
- Hypervisor and a Primary VM controlling test execution.
- SPMC and an SP providing test services.

For the time being only using the indirect messaging test suite, because
it is simple enough to validate the changes made to the hftest framework
and 'primary_with_secondaries' artifacts.

Services are addressed based on their UUID using the newly
added helper functions 'service1/2/3()'. Going forward these functions
shall replace the use of the macros 'SERVICE_VM1/2/3'.

Defined new macro "SERVICE_PARTITION_INFO_GET" to help define the helper
functions service1/2/3 to refer to the service partitions.

This setup is to be extended with more tests, as the hftest framework
and other code gets updated to be compliant with SPs and the SPMC
implementation.

Adding two SPs that have the same uuid as the NWd VMs 'service1' and
'service2'. Added also a third one for future use.

Change-Id: I5325f3b9128b20f856e42e0129450fd90e7d49c3
Signed-off-by: J-Alves <joao.alves@arm.com>