fix: add size to partition_info_get response
The FF-A 1.1 EAC0 specification (Table 13.35) now requires that for
an FFA_SUCCESS response to FFA_PARTITION_INFO_GET the x3/w3 register
contains the size of the partition information descriptors when the
count flag is not set. Implement this in the handler and update the
tests to check this is correctly done.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: I9f1efde183a6cc755c0096c75c43ccf0039b26b2