feat(hftest): primary and secondary VMs are FF-A endpoints
The test setup "primary_with_secondaries" VMs were not configured as
being FF-A endpoints in the hypervisor manifest. As a result, they had a
different load and initialization than those of SPs on top of the SPMC.
This patch makes them FF-A enabled, and they now possess a partition
manifest. The intent is to make secondary VMs as similar as possible to
SPs, such that we reduce discrepancies in instructing Secondary VMs and
SPs from the Primary VM (VM in charge of driving test execution).

The Secondary VMs were also assigned with a different uuid. The UUIDs
are to configure a similar setup using SPs instead of VMs. This
should allow for test partitions (dubbed services) to be referenced
through the uuid and not an ID. UUIDs are generic, whereas FF-A
IDs contain a bit related with the endpoint's world.

QEMU setups have been updated to run with 8 cores to reduce
descripancies with the FVP model run.

Change-Id: Iec356de859be6e586ef1c113e217eb11b29b6162
Signed-off-by: J-Alves <joao.alves@arm.com>