fix: UINTxx_C macro usage
UINTxx_C macro from standard library [1] accepts a scalar/constant value
and concatenates the C standard modifier (U, UL) to make it an unsigned
integer value.
The macro is not meant to be fed with expressions such as masks
e.g. UINT32_C(1 << 3)
The end result is concatenation of the U modifier to the tail of the
expression which is probably not the desired operation.
Most of the code base uses UINTxx_C macro with a single integer value as
parameter, although there are few places where expressions are used.
This change updates the latter to only use constants.

[1] https://git.trustedfirmware.org/hafnium/prebuilts.git/tree/linux-x64/clang/lib64/clang/12.0.5/include/stdint.h#n378

Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Change-Id: I205646171e4b673b6507838f2211efa4c275753c