fix(memory sharing): memory attributes checks
In [1] we enforced that the sender of memory sharing call, should
specify the memory attributes supported by Hafnium. Though this remains
true for a FFA_MEM_SHARE call, this differs to what the spec mandates
for FFA_MEM_DONATE and FFA_MEM_LEND with single borrower. According
to section 10.10.4.2 of FF-A v1.1 EAC0 specification, in those cases the
memory attributes shall remain unspecified.

This patch fixes the checks in the handling of memory sharing calls,
and the tests that were affected by this change.

[1] https://review.trustedfirmware.org/c/hafnium/hafnium/+/13338

Change-Id: I125fbbe4213eb8cfadb53d8fd57ff8e7207dd4e6
Signed-off-by: J-Alves <joao.alves@arm.com>