feat: support for cpu cycle allocation modes
This patch adds initial support for modeling call chains in
SPMC and Normal World scheduled modes.

As per FF-A v1.1 EAC0 section 8.2.3, a call chain represents all SPs
in a sequence of invocations of direct message request. When
execution on a PE is in the secure state, only a single call chain
that runs in the Normal World scheduled mode can exist. However, any
number of call chains that run in the SPMC scheduled mode can exist.

Change-Id: Ic409abd90345d7972e33e42679e53ed8dffb14e7
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>