feat(scheduled mode): start call chain in SPMC scheduled mode
A vCPU becomes the first node in a SPMC scheduled mode call chain
only when:
> It is in waiting state and
> SPMC signals a virtual secure interrupt to it.

SPMC then assigns RTM_SEC_INTERRUPT partition runtime model to this
vCPU.

Execution could be either in normal world or secure world when the
secure interrupt triggers. Both the scenarios are covered by this
patch to implement the actions in response to a secure interrupt.

Change-Id: I0bfd55474b62f1cddd3013371858edd54f6fe84a
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>