feat(scheduled mode): unwind call chain in SPMC scheduled mode
This patch adds support to completely unwind a call chain in SPMC
scheduled mode at the tail end of a FFA_MSG_WAIT ABI invocation.

Note that the last remaining vCPU in a SPMC scheduled call chain
(i.e., the root node) was in WAITING state when a secure virtual
interrupt was signaled to it. It invokes FFA_MSG_WAIT ABI for
signal completion.

Change-Id: I8895e9e9578ccbf5792f57e2701eb20632ff1091
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>