feat(interrupts): bypass blocked intermediate vCPUs
This patch adds the support for the following implementation choice
specified for the SPMC in the Figure 8.13 scenario 1 of FF-A v1.1
EAC0 spec.

The scenario and the implementation choice are as follows:

The secure virtual interrupt is targeted to an vCPU( execution
context of an SP) that ran earlier in the same call chain before
entering the blocked state. The SPMC can signal the virtual interrupt
to the target vCPU (while leaving all the intermediate vCPUs in the
blocked state).

When the target vCPU is done with interrupt handling and invokes
FFA_RUN ABI to resume the vCPU on which it is originally blocked, the
SPMC could instead bypass the intermediate vCPU(s) and directly resume
the vCPU that was originally preempted by the secure interrupt.

This patch also adds another assert to ensure a preempted SP vCPU could
only be a part of NWD scheduled call chain. Spec prohibits an SPMC
scheduled call chain to be preempted by a non secure interrupt.

Change-Id: I1eea4bbec67baf1d59fa1ae2b1478f0bcb361c64
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>