test(memory share): set NS bit in S1 memory mapping
Add function to hftest to fetch memory attributes of the regions mapped
into S1 of the test partitions.
In case sender of memory share is a VM, include NS bit in the memory
attributes before accessing the respective memory region.
This is meant to be a temporary way to detect the use of the NS bit.
As of FF-A v1.1 EAC0, the NS bit should be provided the receiver
through FFA_RETRIEVE_RESP ABI.

Change-Id: I9b700e0059ef479c78859973b6ed7ff7813eb365
Signed-off-by: J-Alves <joao.alves@arm.com>