refactor(tests): exception handlers
The Secondary VMs count the exceptions that ocurred whilst executing
test services. When an exception is triggered, the count of exceptions
is sent to the primary VM. It validates that an exception triggered
when it was actually meant to.

This patch updates the functions that helped with such validation
to use FF-A v1.1 rather than FF-A v1.0 indirect messaging.

Change-Id: I6e1efbb2af4e6f64dbe2914cda6b9bcff2593d73
Signed-off-by: J-Alves <joao.alves@arm.com>