feat(manifest): check SPs don't overlap mem regions
Add check that memory regions defined in the partition manifest
do not overlap.
Since the size of the allocated_fields struct could increase
greatly if MAX_VMs was increased, this patch allocates the number
of entries from the mpool required according to the size of the
allocated_fields struct, helping scaling in the future.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: I5559166c3c94783471bdc3c8bb53f3338c4547a1