fix(memory share): memory attributes on FFA_MEM_RETRIEVE_REQ
Validate that the memory attributes provided in memory region
descriptor when invoking FFA_MEM_RETRIEVE_REQ are those supported by
Hafnium, only if those are specified. If memory attributes are not
specified, skip validation, as the caller expects to receive them in the
FFA_MEM_RETRIEVE_RESP response.

Implemented test to validate that memory can be shared if no memory
type attributes are specified, and that the response contain the
attributes expected to be implemented by hafnium.

Change-Id: Id64489923d8bd58e4ba2a20767b4111f0a4b36ff
Signed-off-by: J-Alves <joao.alves@arm.com>