feat(memory share): multi-receiver from NWd to SWd
Remove restriction of 1 receiver in memory send to the
other world.

This doesn't affect the SPMC as SPs can only use
FFA_MEM_DONATE to the NWd, and there is a check that
for this operation only one receiver can be specified.

This allows the Hypervisor to forward the call to the
SPMC in case multiple receivers have been specified
in the memory transaction descriptor, and at least
one is an SP.

Change-Id: Icfa78bf73a992e7127d627b739172285a6648020
Signed-off-by: J-Alves <joao.alves@arm.com>