feat(memory share): share memory with the other world
The Hypervisor keeps its implementation, forwarding memory share to
the SPMC, which shall validate the request against the 'other_world_vm'
configured memory permissions, track the memory operation in its internal
'share_state' variables, allocate a handle and return the result to the
Hypervisor.

The SPMC doesn't forward the request to the hypervisor. It validates
the memory sharing operation internally, and returns the handle to the
caller endpoint.

Created 'ffa_memory_internal.h' header to declare private functions and
symbols to the ffa_memory module. With this patch, some functionality
was made hypervisor specific. Creating this header, allows for the
hypervisor implementation to reference functions it needs, whilst
seperating declaration from ffa_memory module's public interface used in
'api.c'.

Change-Id: Ia9c14cf112f0143869ac4a7ba9b495b08e38f9bb
Signed-off-by: J-Alves <joao.alves@arm.com>