refactor(memory share): memory retrieve from hypervisor
According to section 16.4.3 of the FF-A 1.1 EAC0 specification
the hypervisor may issue an FFA_MEM_RETRIEVE_REQ with memory region
descriptor specifying the handle of the transaction, and the remaining
descriptor zeroed. To obtain the information of an on going memory
sharing operation involving the SWd.

This was being done in the handling of FFA_MEM_RECLAIM when the
memory handle has been previously allocated by the SPMC. This
patch factors out the code that issues the retrieve request to the
SPMC from the handling of FFA_MEM_RECLAIM, and places it in its
own function. The aim is to better organize the code for reclaiming
memory regions from SPMC.

Change-Id: I030a7dc078fe08a539b1dec85341cfe067882ec9
Signed-off-by: J-Alves <joao.alves@arm.com>