refactor: build option for interrupt ids
Make HF_NUM_INTIDS a build option for platform with large number of
interrupts. For large systems with GIC extensions that allow > 4K
interrupts, 64 interrupts is inadequate. An example partition that may
require more than 64 interrupts is a RAS SP that handles many interrupts
from many sources on a large server type SoC.
Note that the limit of max number of interrupt IDs is set to 5120 based
on the latest GIC specs for extended SPI range (not using ITS).

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>
Change-Id: Ic3b1741b61a4a796da36f505357af86949b249e6