fix(init): Remove defrag on hypervisor tables
The init function calls mm_defrag on the hypervisor or spmc tables. This
is problematic because the defrag functions unmaps finer grained entries
and replace it with block entries, which requires the break-before-make
approach where the entry is first replaced with an invalid entry,
invalidate tlbs and then replace with the new coarse grained entry. When
this replacement occurs on the hypervisor/spmcs code or stack pages
itself, for example, the defrag function unmaps the pages of stack that
it is executing on currently on, the defrag function will take an
exception and crash the hypervisor/spmc. This crash is highly dependent
on the layout of the spmc/hypervisor and especially when hafnium is
scaled to large number of pcpu, vcpu and stacks, since that would mean
mapping and unmapping a lot more pages in the hypervisor and more
opportunity to defrag the hypervisor or stack pages.
This patch removes the defrag call on the hypervisor/spmc tables to
avoid this issue. Note that it is removed in other code paths besides
just the init function, since the same problem applies.

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>
Change-Id: I7ca41d30fbe3553d26b754985f9524dbc1e25693