fix(memory sharing): test donate_vms
In a setup with Hyp and SPMC the test was causing problems due to a
current implementation limitation: The SPMC infers the security of the
memory based on the sender's ID. This is a flawed approach, that helped
validate few cases and should change in the future.

The test's purpose was to validate service partitions could donate
memory among them. Prior, the memory donated was always NS, and was
consecutively donated from NWd prim into service1 and from service1 to
service2. The change was to avoid doing the first donation, and donate
memory statically allocated within service1. This would avoid any
issues regarding the memory's security state.

Change-Id: I73dec48e1b46571e89bdca173f0334febd8b2f62
Signed-off-by: J-Alves <joao.alves@arm.com>