feat(tests): helper to check occurrence of exception
In some of the tests the primary VM looks for the occurrence of
exceptions in the service partition. This can be done in two different
ways, depending on the exception level of the service:
- EL1 keeps count of the number of exceptions it handled, and sends it
in a message to the Primary VM.
- EL0 is aborted, and the SPMC/Hyp returns execution to the primary VM
with FFA_ERROR and error code FFA_ABORTED.

The newly added helper function is to be used by the primary VM to
check for the occurrence of exceptions, and works for both cases.

Change-Id: I5db9da013f41bf52bb3f7d069b65283191921de7
Signed-off-by: J-Alves <joao.alves@arm.com>