refactor: check return values
Add CHECK statements on return values during mmu initialization. it is
possible that memory allocation fails if the heap is too small, which
will only be caught via indirect faults.
This patch also fixes the function definition for the fake
arch_mm_stack_init that returned void, but should return a bool, and
changes the fake layout so that start address in the layout is not 0,
which is the same as NULL and causes test failures.

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>
Change-Id: I6ef0b6e31f9b5bfd683b18d43c21019970685dd9