fix(ffa_version): set ff-a version of hypervisor
The hypervisor calls ffa_version at initialization, which should set its
associated version in the SPMC.
This patch forces hypervisor's version to be the same as the SPMC's for
the sake of enabling backward compatibility patches.
This indicates there is presumably an issue with SPMD implementation
which needs fixing. After doing that, this change should be dropped.

Change-Id: If11a2d96b84545bc61ef936635c371fe7ef49246
Signed-off-by: J-Alves <joao.alves@arm.com>