fix: make the invocation of ffa_msg_wait v1.1 compliant
The invocation of FFA_MSG_WAIT at secure virtual FF-A instance is
made to be compliant with FF-A v1.1 EAC0 specification. It's
only use is to perform the state transition from RUNNING to WAITING
for the following Partition runtime models: RTM_FFA_RUN,
RTM_SEC_INTERRUPT, and RTM_SP_INIT.

However, the invocation of FFA_MSG_WAIT at non-secure virtual FF-A
instance is made to be compliant with the FF-A v1.0 spec for
legacy reasons.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: Ic63f07c9ce6e6c3be9df2a849900b30eb61fcf6d