fix(memory sharing): remove ffa_msg_wait calls
FFA_MSG_WAIT was updated to match FF-A v1.1 EAC0 specification [1].
The SPMC puts the calling SP's vCPU in waiting state. Whereas
for VMs the FFA_MSG_WAIT call returns in case there is a pending
message in the RX buffer.

The memory sharing tests had a call to FFA_MSG_WAIT to ensure they
had a message with the memory transaction descriptor at the
beginning of the test services. Those were not needed, given the
tests were always sending a message, which is also validated where
needed using the FFA_NOTIFICATION_GET interface to retrieve framework
notifications.

In addition, with the update to FFA_MSG_WAIT the execution of
memory sharing tests breaks, when using it with SPs. This
patch deletes all calls to FFA_MSG_WAIT at the beginning of test
services.

[1] https://review.trustedfirmware.org/c/hafnium/hafnium/+/17406

Change-Id: Ifebec23d9dd984d7af10bfb2a42e530aabbeef00
Signed-off-by: J-Alves <joao.alves@arm.com>