feat(ff-a): report execution state in partition info get
FF-A Version 1.1 adds several additional flags to
the properties field of the partition information
descriptors. Add flag for execution state and test
that it is set when FF-A v1.1 is specified and zero
when FF-A v1.0 descriptors are being used.

Currently only 64-bit execution state is supported
for FF-A compliant partitions.

Signed-off-by: Kathleen Capella <kathleen.capella@arm.com>
Change-Id: Ic0e406bd491b5bae248704ba81ed5620e777e26b