feat(memory sharing): clear memory after relinquish
The FFA_MEM_RETRIEVE_REQ_32 interface may receive a
flag to request the SPMC to clear memory upon
memory relinquish.

This patch adds support for this functionality:
- Added flag to capture aforementioned flag is set by one
of the borrowers, in the 'struct share_state'.
- The handling of FFA_MEM_RELINQUISH looks for this flag,
checks that all remaining borrowers relinquished access,
and clears memory.
- To clear memory, the SPMC first maps the memory into
its address space, and then writes 0s to the memory. In
doing so, uses the owner of the memory region to determine
the NS bit configuration.

Change-Id: I8eb5bc10e1488ab61ac4ef9ccb27f4470ef50b77
Signed-off-by: J-Alves <joao.alves@arm.com>