ci: update prebuilts to qemu7.1.0
Update prebuilts submodule refreshing the x86 qemu prebuilt binary and
introducing the aarch64 qemu prebuilt binary.

Note: the qemu machine is forced to virt-6.2 to prevent enabling the
FEAT_LPA2 feature. The reason is the use in the Hafnium CI of linux
kernel 5.10 which has a bug preventing it to boot properly when
FEAT_LPA2 is implemented [1].

[1] https://github.com/qemu/qemu/commit/0942820408dc788560f6968e9b5f011803b846c2

Change-Id: I1aca0140bcbc0fe8d053c673e9df9547cd51a36a
Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>