feat(ff-a): send retrieve request helpers
Defined helpers to aid with sending a retrieve
request message from the lender to the borrower.
The wrappers initialize a message with retrieve
request memory region, and send it as a message to
the borrower.

In memory sharing tests the borrower copies the
message from its RX buffer into its TX buffer,
and invokes FFA_MEM_RETRIEVE_REQ.

Change-Id: I983bff0f560edc61e977d4b902f23ac8a121a1b3
Signed-off-by: J-Alves <joao.alves@arm.com>