refactor(memory sharing): skip tests
Patch [1] skipped some memory sharing tests, in which
there is memory shared from service partition to
the primary VM. The tests are skipped due to a
known limitation, on the way we determine the NS bit
state of the memory being shared.

Patch [2] added a more efficient way to skip
tests.

It seems waistful to launch a setup running a test
to then skip its execution.

This patch uses [2] to clean up after [1].

Tests from the memory sharing test suite won't execute
if the precondition function returns false for a given
testing setup.

[1] https://review.trustedfirmware.org/c/hafnium/hafnium/+/16525
[2] https://review.trustedfirmware.org/c/hafnium/hafnium/+/18230

Change-Id: I03dd9cf5059221db7207ea9c4173ceb136dc7d7c
Signed-off-by: J-Alves <joao.alves@arm.com>