ci(memory sharing): skip redundant tests
Currently the memory sharing tests are part of two setups:
1- PVM + SVMs + Hypervisor
2- PVM + HYP + SPs + SPMC

The tests validating sanity checks to sender's calls to
share, lend or donate ABIs have equivalent flow. They
are contained to the NWd. That said, having them repeated
in setup 2 is redundant. Also considering setup 2 is
currently only executing in FVP, and it takes a long time to
launch a platform with the testing setup, we could skip
these tests in setup 2.

This patch adds a precondition function that checks for
presence of secure FF-A endpoints. If it does, it returns true.
This precondition function can be used to skip tests
using the newly introduced macro TEST_PRECONDITION, to skip
the aforementioned redundant tests.

Change-Id: I7bc77a4a062c7309c9c87ca0a05da9bc97bdf337
Signed-off-by: J-Alves <joao.alves@arm.com>