fix: rxtx unmap forwarding and handling
For calls to FFA_RXTX_UNMAP at the non-secure physical instance
the FF-A v1.1 prescribes that x1 contains the ID of the
NS-endpoint, in bits [31:16].

In forwarding the call, the hypervisor was setting the arguments
correctly, but the SPMC would retrieve without shifting the
register with arguments. This patch fixes that.

Also, this patch makes it such the Hypervisor only
forwards the call to the SPMC if the FF-A version of
the caller is FF-A v1.1.

Change-Id: I7565cc52d387c550295114ab79b2e5bb76b4281e
Signed-off-by: J-Alves <joao.alves@arm.com>