fix(ff-a): versioned partition info get test
Revert the change introduced in [1] to the tests, which was
a workaround to the real issue.

In addition, fixed the original problem. The test was:
1- Setting FF-A version to v1.0, and checking the FF-A v1.0
partition info get descriptors.
2- Setting FF-A version to v1.1, and checking the FF-A v1.1
partition info get descriptors.

In between there was no remapping of RXTX buffers. As a result,
the SPMC didn't have mapped the buffers of the caller to
FFA_PARTITION_INFO_GET, which for v1.1 partitions calls
FFA_RX_ACQUIRE for the hypervisor to own the RX buffer,
and avoid race conditions with the SPMC. So the FFA_RX_ACQUIRE,
would fail and as a result would also the FFA_PARTITION_INFO_GET.

This patch changes the referred test to only validate the return to
FFA_PARTITION_INFO_GET for a v1.0 partition. The change of FF-A version
in runtime should be limited to a single call. Such change in hafnium
shall be enforced in the future. This patch only cares about fixing the
issue with handling RXTX map for the time being.

[1] https://review.trustedfirmware.org/c/hafnium/hafnium/+/14472

Change-Id: Idca2e18b26dd2a58b8b6ea2f586be6008d1022b8
Signed-off-by: J-Alves <joao.alves@arm.com>