refactor: avoid unnecessary lock/unlock on rxtx map
The handling of FFA_RXTX_MAP ABI was doing an unnecessary
lock/unlock to the lock guarding the caller vm.

Having a look it seems these locks/unlock calls would be
useful when the caller is from NWd into the SPMC.
This patch does some small refactoring to avoid the
unnecessary lock acquisition.

Change-Id: I609b67830b588e095653803f79f6a5329ba5f4fc
Signed-off-by: J-Alves <joao.alves@arm.com>