refactor: helpers for partition info get in the tests
Planning to add the ability to unmap the rxtx buffers, in the test
partitions and change the implementation of 'set_up_mailbox'.
This patch helps to prepare for such change.

Currently 'set_up_mailbox' is used everywhere, and internally statically
determines if the buffers have been mapped. If so, return the pair
addresses that were configuraed by the initial call to FFA_RXTX_MAP ABI.
As there is the need to be able to unmap these addresses, this
optimization needs to be undone.

To prepare for that, all helper functions to retrieve service partitions
IDs need to be refactored, to receive the address of the RX buffer.

So 'set_up_mailbox' function should be used once to map RXTX buffers,
and from that moment, pass their addresses to functions that may need
it.

The precondition are invoked from 'test/hftest/common.c' which is linked
with partition targets, as well as 'hypervisor' targets. On the latter,
it doesn't link with any partition FF-A related code. So to help with
using mailbox within the functions, and maintaining code isolation at
buildsystem, created a static helper function for the precondition
functions of 'primary_with_secondaries'.

Change-Id: If0b04c89ffb7597cf1e237ddedc8df75e102d96c
Signed-off-by: J-Alves <joao.alves@arm.com>